+++ changelog.md
@@ -41,6 +41,7 @@ graph.
 files.
 * added check for popup blocking.
 * new function that allows to center on the node with more links
+* added new button to center on the node being displayed in the popup
 
 ## Version 1.4.1